Add checker for magenta mutex
The checker analyzes uses of mutex_init, mutex_acquire,
mutex_release, and mutex_destroy. It performs the core checks
on usage of these APIs on mutex to make sure they are correct.
For instance, we want to make sure a mutexes are always initialized
in the constructor and destroyed in the destructor. Also, we do not
want to have paths where a mutex is released while not have been
acquired first.

10 test files are also added to the Clang checkers' test suit.

Change-Id: Ie96456f0a01246467c6fc3311023a270324d76d1